# Faster version of the data generator