@gunnarmorling 2024-01-03
Nice one! Has that code in the `org.schwietzke` package been written by you? Or is it taken from somewhere else?