@rschwietzke 2024-01-03
> Nice one! Has that code in the `org.schwietzke` package been written by you? Or is it taken from somewhere else?
All my code, which has been already contributed to [XLT](https://github.com/Xceptance/XLT/) and [HtmlUnit Neko](https://github.com/HtmlUnit/htmlunit-neko). 

`FastRandom` is inspired by a post somewhere and some demo code, don't recall it anymore. Poor randomness, but good enough for our demo case and repeatable if needed. There is also a more elaborate version here: https://github.com/Xceptance/XLT/blob/develop/src/main/java/it/unimi/dsi/util/FastRandom.java which is based on http://prng.di.unimi.it/. 

The `CheapCharBuffer ` is too big for our case but taken from my last PR to HtmlUnit Neko: https://github.com/HtmlUnit/htmlunit-neko/blob/master/src/main/java/org/htmlunit/cyberneko/xerces/xni/XMLString.java Has a different name for compatibility reasons there.